pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "winget"
version = "0.0.1"
description = "A library for interacting with the Windows Package Manager."
readme = "README.md"
authors = [{ name = "Emil Kloeden", email = "emilkloeden@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["winget"]
dependencies = [
    "requests == 2.28.1",
    "PyYAML == 6.0",
    'importlib_resources; python_version < "3.10"',
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["black", "bumpver", "isort", "pip-tools", "pytest"]

[project.urls]
Github = "https://github.com/emilkloeden/winget"
